requirements.txt
asgiref==3.7.2
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-reversion==5.0.5
django-seed==0.3.1
django-storages==1.14.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-flex-fields==1.0.2
Faker==19.8.0
flake8==6.1.0
fpdf==1.7.2
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
Pillow==10.0.1
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
reversion==0.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
